ocr-service/requirements.txt
easyocr
pika
boto3
torch
opencv-python-headless
python-dotenv
numpy
requests